Remove ModuleFile from header, move ModuleDeps implementation
